moved date formating into eventcard
