+++ src/components/EventCard.js
@@ -1,45 +1,72 @@
 // @flow
 import React from "react";
-import { View, Text, StyleSheet } from "react-native";
+import { View, StyleSheet } from "react-native";
+import isSameDay from "date-fns/is_same_day";
+import formatDate from "date-fns/format";
 import {
   priceTagBgColor,
   imageBgColor,
   cardBgColor,
   textColor
 } from "../constants/colors";
+import Text from "./Text";
 
 type Props = {
   name: string,
   locationName: string,
   price: string,
-  startTime: integer
+  eventStartTime: integer,
+  eventEndTime: integer
 };
+const removeTimezoneFromDateString = isoString => isoString.slice(0, -6);
 
-const EventCard = ({ name, locationName, startTime, price }: Props) => (
-  <View style={styles.eventCard}>
-    <View style={styles.imageContainer}>
-      <View style={styles.priceTagContainer}>
-        <Text style={styles.eventPrice}>{price}</Text>
+const EventCard = ({
+  name,
+  locationName,
+  eventStartTime,
+  eventEndTime,
+  price
+}: Props) => {
+  const startTime = removeTimezoneFromDateString(eventStartTime);
+  const endTime = removeTimezoneFromDateString(eventEndTime);
+  const dateFormat = "DD MMMM YYYY";
+  const timeFormat = "HH:mm";
+  const dateDisplay = isSameDay(startTime, endTime)
+    ? formatDate(startTime, dateFormat)
+    : `${formatDate(startTime, dateFormat)} - ${formatDate(
+        endTime,
+        dateFormat
+      )}`;
+  const timeDisplay = `${formatDate(startTime, timeFormat)} - ${formatDate(
+    endTime,
+    timeFormat
+  )}`;
+  return (
+    <View style={styles.eventCard}>
+      <View style={styles.imageContainer}>
+        <View style={styles.priceTagContainer}>
+          <Text style={styles.eventPrice}>{price}</Text>
+        </View>
       </View>
-    </View>
-    <View style={styles.eventCardDetails}>
-      <Text ellipsizeMode="tail" numberOfLines={1} style={styles.eventTime}>
-        {startTime}
-      </Text>
-      <View style={styles.eventNameContainer}>
-        <Text ellipsizeMode="tail" numberOfLines={1} style={styles.eventName}>
-          {name}
+      <View style={styles.eventCardDetails}>
+        <Text ellipsizeMode="tail" style={styles.eventTime}>
+          {timeDisplay}
+        </Text>
+        <View style={styles.eventNameContainer}>
+          <Text ellipsizeMode="tail" style={styles.eventName}>
+            {name}
+          </Text>
+        </View>
+        <Text ellipsizeMode="tail" style={styles.eventLocation}>
+          {locationName}
         </Text>
       </View>
-      <Text ellipsizeMode="tail" numberOfLines={1} style={styles.eventLocation}>
-        {locationName}
-      </Text>
+      <View style={styles.heartButtonContainer}>
+        <Text style={styles.heartIcon}>heart</Text>
+      </View>
     </View>
-    <View style={styles.heartButtonContainer}>
-      <Text style={styles.heartIcon}>heart</Text>
-    </View>
-  </View>
-);
+  );
+};
 
 const styles = StyleSheet.create({
   eventCard: {
