--- src/components/EventLocation.js
@@ -1,6 +1,0 @@
-import React from "react";
-import { Text } from "react-native";
-
-const EventLocation = ({ location }: Props) => <Text>{location}</Text>;
-
-export default EventLocation;
