--- src/components/EventTime.js
@@ -1,6 +1,0 @@
-import React from "react";
-import { Text } from "react-native";
-
-const EventTime = () => <Text>time</Text>;
-
-export default EventTime;
